refactor: move util functions into ./src/utils
